requirements.txt
Flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
Flask-SQLAlchemy>=3.1.1
PyMySQL>=1.1.0